[fixed] cleaned up some of the _this.get calls [fixed] some more coverity defects

+++ MMOCoreORB/src/server/zone/objects/player/PlayerObjectImplementation.cpp
@@ -1608,7 +1608,7 @@ void PlayerObjectImplementation::activateForcePowerRegen() {
 	if (forceRegenerationEvent == NULL) {
 		forceRegenerationEvent = new ForceRegenerationEvent(_this.getReferenceUnsafeStaticCast());
 
-		float timer = getForcePowerRegen() / 5;
+		float timer = ((float) getForcePowerRegen()) / 5.f;
 		float scheduledTime = 10 / timer;
 		uint64 miliTime = static_cast<uint64>(scheduledTime * 1000.f);
 		forceRegenerationEvent->schedule(miliTime);
